moved grotto functions into a Main subfolder
