+++ include/Grotto/Main/TreasureMapMetadata.h
